fix: rename prisma to prismaclient.js
